backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
supabase==2.8.0
pdfplumber==0.11.4
sentence-transformers==3.1.1
google-generativeai==0.8.3
numpy==2.1.1
python-dotenv==1.0.1

